LPS-98077 Better naming
