+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/handlers/fieldAddedHandler.es.js
@@ -15,7 +15,7 @@
 import {FormSupport} from 'dynamic-data-mapping-form-renderer';
 
 import {createField} from '../../../util/fieldSupport.es';
-import {updateFocusedField} from '../util/focusedField.es';
+import {updateFocusedField} from '../util/settingsContext.es';
 
 const getContext = (context, nestedIndexes = []) => {
 	if (nestedIndexes.length) {
